# HW2_GitProxy/openapi.yaml
openapi: 3.1.0
info:
  title: GitHub REST API proxy gateway server
  version: 1.0.0
  description: |
    Gateway API to manage GitHub Issues on a configured repository, and receive GitHub webhooks.
servers:
  - url: http://localhost:{port}
    variables:
      port:
        default: "5000"
paths:
  /issues:
    post:
      summary: Create an issue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIssueRequest'
            examples:
              create:
                value:
                  title: "Bug: cannot save profile"
                  body: "Steps to reproduce..."
                  labels: [bug, urgent]
      responses:
        '201':
          description: Created
          headers:
            Location:
              description: Location of the created resource
              schema: { type: string, example: "/issues/123" }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    get:
      summary: List issues
      parameters:
        - in: query
          name: state
          schema: { type: string, enum: [open, closed, all], default: open }
        - in: query
          name: labels
          schema: { type: string }
          description: Comma-separated labels
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: per_page
          schema: { type: integer, minimum: 1, maximum: 100, default: 30 }
      responses:
        '200':
          description: List of issues
          headers:
            Link:
              description: Pagination links as per GitHub semantics
              schema: { type: string }
            X-RateLimit-Limit:
              schema: { type: integer }
            X-RateLimit-Remaining:
              schema: { type: integer }
            X-RateLimit-Reset:
              schema: { type: integer }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Issue' }
...
components:
  schemas:
    Label:
      type: object
      properties:
        name: { type: string }
    Issue:
      type: object
      required: [number, html_url, state, title, created_at, updated_at]
      properties:
        number: { type: integer }
        html_url: { type: string, format: uri }
        state: { type: string, enum: [open, closed] }
        title: { type: string }
        body: { type: string, nullable: true }
        labels:
          type: array
          items: { $ref: '#/components/schemas/Label' }
        created_at: { type: string }
        updated_at: { type: string }
...
    CreateIssueRequest:
      type: object
      required: [title]
      properties:
        title: { type: string }
        body: { type: string }
        labels:
          type: array
          items: { type: string }
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
        details: { type: object }
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  securitySchemes:
    githubToken:
      type: http
      scheme: bearer
      bearerFormat: GitHub PAT
security:
  - {}